API Call for internal call added.
