--- server/index.jsx
@@ -1,13 +1,0 @@
-const express = require('express');
-const bodyParser = require('body-parser');
-const db = require('../database/index.js');
-
-const app = express();
-
-app.use(bodyParser.json());
-app.use(bodyParser.urlencoded({ extended: true }));
-app.use(express.static(__dirname + '/../dist'));
-
-const PORT = 3102;
-
-app.listen(PORT, () => console.log('listening on port', PORT));
