+++ src/index.jsx
@@ -6,7 +6,7 @@ class Navbar extends React.Component {
   constructor(props) {
     super(props);
     this.state = {
-      //
+      countdown: 14,
     }
   }
